AirBooks-1.0.8
This is the final release for the core functions:

- The application is checking the location every day at 11:55 pm and
perform the following IF scenarios:
* IF the user is inside 75 miles radius from the home airport do nothing
* IF the user is outside the 75 miles radius from the home airport get
the location and save a record.
* IF app is closed, launch the app and get the location ( due to GPS
restrictions from Google, the only working option (up to Now) is to
launch the app at 11:55 in order to obtain the location by the best
provider. It could be a little annoying to the user (if using the phone
at that time, the app will pop up or next time the user will use the
phone the app will be open.
* If the device was rebooted, maintain the alarm to check location at
11:55 pm

- Find the nearest airport by comparing the distance between the current
coordinates and the airports coordinates, once the closest airport is
located, it will take that airport's city name and request the perDiem
value for that city.