Version 1.0.10
Changed setting of Country, State, City in MyLocationManager to one method.  No need to make three seperate calls to the 
geolocator when the same inforamtion is used each time.  Also added a check to keep the app from crashing when getting these values if it fails to return an address for some reason.

Commented out the LocationManager class since it was no longer being used and was replaced by MyLocationManager.

Added swiping functionality to MyLocationManager, ViewTripHistory, TaxDeductionsToDate, and EditUserInfo.  User can now swipe left/right to navigate through these screens. (Currently the app still opens to the main page.  This will eventually be changed to open straight to the 'Current Trip' class when it exists.  Also some of the code on the main page needs to be moved to the Current Trip class.)